feat: Update Photo III

+++ src/controllers/PhotoC.js
@@ -1,4 +1,6 @@
 import multer from 'multer';
+import { resolve } from 'path';
+import fs from 'fs';
 
 import multerConfig from '../config/multer';
 import Photo from '../models/PhotoM';
@@ -14,13 +16,29 @@ class PhotoController {
         });
       }
 
+      if (!req.file) {
+        return res.status(400).json({
+          errors: ['Missing file.'],
+        });
+      }
+
       try {
         const { originalname, filename } = req.file;
         const { student_id } = req.body;
 
         const photo = await Photo.findAll({ where: { student_id } });
+        const path = resolve(__dirname, '..', '..', 'uploads', 'images', photo[0].filename);
 
         if (photo.length !== 0) {
+          fs.unlink(path, (err) => {
+            if (err) {
+              return res.status(400).json({
+                errors: [err.message],
+              });
+            }
+            return true;
+          });
+
           await Photo.destroy({ where: { student_id } });
         }
